backjar/requirements.txt
fastapi>=0.111.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.35.0
pywebpush>=2.0.0,<3.0.0
firebase-admin>=6.5.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.7.0,<3.0.0
httpx>=0.27.0,<0.29.0
apscheduler>=3.10.0,<4.0.0